requirements.txt
openai>=1.11.1
pandas
plotly
plotly-calplot
selenium
tenacity
python-dotenv